Front-end almost finished

+++ capstone/java/database/schema.sql
@@ -55,7 +55,7 @@ CREATE SEQUENCE seq_activity_id
 CREATE TABLE reading_activity (
     activity_id int NOT NULL DEFAULT nextval('seq_activity_id'),
     user_id int NOT NULL,
-    isbn int NOT NULL,
+    isbn numeric NOT NULL,
     minutes_read int DEFAULT 0,
     date_read timestamp,
     notes varchar (250),
